+++ quadkey/bq/sql/BBOX.sql
@@ -11,6 +11,5 @@ CREATE OR REPLACE FUNCTION `@@BQ_PROJECTID@@.@@BQ_DATASET_QUADKEY@@.BBOX`
     LANGUAGE js
     OPTIONS (library=["@@QUADKEY_BQ_LIBRARY@@"])
 AS """
-    var b = bbox(quadint);  
-    return [b.min.lng, b.min.lat, b.max.lng, b.max.lat];
+    return bbox(quadint);
 """;
