+++ quadkey/bq/sql/KRING.sql
@@ -11,14 +11,14 @@ CREATE OR REPLACE FUNCTION `@@BQ_PROJECTID@@.@@BQ_DATASET_QUADKEY@@.KRING`
     LANGUAGE js
     OPTIONS (library=["@@QUADKEY_BQ_LIBRARY@@"])
 AS """
-    var left      = sibling(quadint,'left').toString();
-    var topleft   = sibling(left,'up').toString();
-    var downleft  = sibling(left,'down').toString();
-    var right     = sibling(quadint,'right').toString();
-    var topright  = sibling(right,'up').toString();
-    var downright = sibling(right,'down').toString();
-    var up        = sibling(quadint,'up').toString();
-    var down      = sibling(quadint,'down').toString();
+    let left      = sibling(quadint,'left').toString();
+    let topleft   = sibling(left,'up').toString();
+    let downleft  = sibling(left,'down').toString();
+    let right     = sibling(quadint,'right').toString();
+    let topright  = sibling(right,'up').toString();
+    let downright = sibling(right,'down').toString();
+    let up        = sibling(quadint,'up').toString();
+    let down      = sibling(quadint,'down').toString();
 
     return [left, topleft, downleft, right, topright, downright, up, down, quadint];
 """;
