+++ quadkey/bq/sql/ST_ASQUADINT.sql
@@ -11,7 +11,7 @@ CREATE OR REPLACE FUNCTION `@@BQ_PROJECTID@@.@@BQ_DATASET_QUADKEY@@.LONGLAT_ASQU
     LANGUAGE js
     OPTIONS (library=["@@QUADKEY_BQ_LIBRARY@@"])
 AS """
-    return quadintFromLocation({lng: longitude, lat: latitude}, resolution).toString();
+    return quadintFromLocation(longitude, latitude, resolution).toString();
 """;
 
 CREATE OR REPLACE FUNCTION `@@BQ_PROJECTID@@.@@BQ_DATASET_QUADKEY@@.ST_ASQUADINT`
