+++ quadkey/library/package.json
@@ -11,5 +11,10 @@
     "eslint-plugin-promise": "^4.2.1"
   },
   "author": "",
-  "license": "ISC"
+  "license": "ISC",
+  "dependencies": {
+    "@mapbox/tile-cover": "^3.0.2",
+    "@mapbox/tilebelt": "^1.0.2",
+    "browserify": "^17.0.0"
+  }
 }
